[checker] Move dependency checker code into index
