+++ samlang-core/checker/__tests/index.test.ts
@@ -1,4 +1,5 @@
 import {
+  DependencyTracker,
   typeCheckSources,
   typeCheckSourcesIncrementally,
   // eslint-disable-next-line camelcase
@@ -9,6 +10,59 @@ import { ModuleReference } from 'samlang-core-ast/common-nodes';
 import { createGlobalErrorCollector } from 'samlang-core-errors';
 import { parseSamlangModuleFromText } from 'samlang-core-parser';
 import { mapOf, hashMapOf } from 'samlang-core-utils';
+
+it('can track and update dependencies', () => {
+  const tracker = new DependencyTracker();
+  const moduleA = new ModuleReference(['A']);
+  const moduleB = new ModuleReference(['B']);
+  const moduleC = new ModuleReference(['C']);
+  const moduleD = new ModuleReference(['D']);
+  const moduleE = new ModuleReference(['E']);
+
+  // Wave 1
+  tracker.update(moduleA, [moduleB, moduleC]);
+  tracker.update(moduleD, [moduleB, moduleC]);
+  tracker.update(moduleE, [moduleB, moduleC]);
+  tracker.update(moduleA, [moduleB, moduleC]);
+  tracker.update(moduleD, [moduleB, moduleC]);
+  tracker.update(moduleE, [moduleB, moduleC]);
+  expect(tracker.getForwardDependencies(moduleA).toArray()).toEqual([moduleB, moduleC]);
+  expect(tracker.getForwardDependencies(moduleB).toArray()).toEqual([]);
+  expect(tracker.getForwardDependencies(moduleC).toArray()).toEqual([]);
+  expect(tracker.getForwardDependencies(moduleD).toArray()).toEqual([moduleB, moduleC]);
+  expect(tracker.getForwardDependencies(moduleE).toArray()).toEqual([moduleB, moduleC]);
+  expect(tracker.getReverseDependencies(moduleA).toArray()).toEqual([]);
+  expect(tracker.getReverseDependencies(moduleB).toArray()).toEqual([moduleA, moduleD, moduleE]);
+  expect(tracker.getReverseDependencies(moduleC).toArray()).toEqual([moduleA, moduleD, moduleE]);
+  expect(tracker.getReverseDependencies(moduleD).toArray()).toEqual([]);
+  expect(tracker.getReverseDependencies(moduleE).toArray()).toEqual([]);
+
+  // Wave 2
+  tracker.update(moduleA, [moduleD, moduleE]);
+  expect(tracker.getForwardDependencies(moduleA).toArray()).toEqual([moduleD, moduleE]);
+  expect(tracker.getForwardDependencies(moduleB).toArray()).toEqual([]);
+  expect(tracker.getForwardDependencies(moduleC).toArray()).toEqual([]);
+  expect(tracker.getForwardDependencies(moduleD).toArray()).toEqual([moduleB, moduleC]);
+  expect(tracker.getForwardDependencies(moduleE).toArray()).toEqual([moduleB, moduleC]);
+  expect(tracker.getReverseDependencies(moduleA).toArray()).toEqual([]);
+  expect(tracker.getReverseDependencies(moduleB).toArray()).toEqual([moduleD, moduleE]);
+  expect(tracker.getReverseDependencies(moduleC).toArray()).toEqual([moduleD, moduleE]);
+  expect(tracker.getReverseDependencies(moduleD).toArray()).toEqual([moduleA]);
+  expect(tracker.getReverseDependencies(moduleE).toArray()).toEqual([moduleA]);
+
+  // Wave 3
+  tracker.update(moduleA);
+  expect(tracker.getForwardDependencies(moduleA).toArray()).toEqual([]);
+  expect(tracker.getForwardDependencies(moduleB).toArray()).toEqual([]);
+  expect(tracker.getForwardDependencies(moduleC).toArray()).toEqual([]);
+  expect(tracker.getForwardDependencies(moduleD).toArray()).toEqual([moduleB, moduleC]);
+  expect(tracker.getForwardDependencies(moduleE).toArray()).toEqual([moduleB, moduleC]);
+  expect(tracker.getReverseDependencies(moduleA).toArray()).toEqual([]);
+  expect(tracker.getReverseDependencies(moduleB).toArray()).toEqual([moduleD, moduleE]);
+  expect(tracker.getReverseDependencies(moduleC).toArray()).toEqual([moduleD, moduleE]);
+  expect(tracker.getReverseDependencies(moduleD).toArray()).toEqual([]);
+  expect(tracker.getReverseDependencies(moduleE).toArray()).toEqual([]);
+});
 
 it('typeCheckSources integration smoke test (passing case)', () => {
   const sourceA = `class A { function a(): int = 42 }`;
